+++ app/components/wallet-overview/index.tsx
@@ -1,1 +1,1 @@
-export * from "./wallet-overview-redesigned"
+export * from "./wallet-overview"
