+++ app/rne-theme/themed.d.ts
@@ -29,6 +29,7 @@ declare module "@rneui/themed" {
     grey7: string
     grey8: string
     grey10: string
+    grey9: string
     horizonBlue: {
       colors: string[]
       start: {
